fix: rename, correct according to comments  bb-125
