--- apps/mobile/src/libs/components/custom-radio-group/libs/constants/constants.ts
@@ -1,5 +1,0 @@
-const DEFAULT_BORDER_SIZE = 0;
-const DEFAULT_ACTIVE_BORDER_SIZE = 4;
-const DEFAULT_CIRCLE_SIZE = 16;
-
-export { DEFAULT_ACTIVE_BORDER_SIZE, DEFAULT_BORDER_SIZE, DEFAULT_CIRCLE_SIZE };
