+++ apps/mobile/src/libs/components/radio-group/libs/constants/constants.ts
@@ -1,0 +1,1 @@
+export { DefaultStyles } from "./default-styles";
