+++ apps/mobile/src/libs/components/radio-group/libs/constants/default-styles.ts
@@ -1,0 +1,7 @@
+const DefaultStyles = {
+	DEFAULT_ACTIVE_BORDER_SIZE: 4,
+	DEFAULT_BORDER_SIZE: 0,
+	DEFAULT_CIRCLE_SIZE: 16,
+};
+
+export { DefaultStyles };
