+++ apps/mobile/src/screens/welcome-screen/welcome-screen.tsx
@@ -2,19 +2,19 @@ import React from "react";
 
 import { RadioGroup, Text, View } from "~/libs/components/components";
 
+const mockData = [
+	{ label: "Family and relationships", value: "family" },
+	{ label: "Career and professional growth", value: "career" },
+	{ label: "Personal development and learning", value: "personal" },
+	{ label: "Health and well-being", value: "health" },
+	{ label: "Financial stability", value: "financial" },
+];
+
 const WelcomeScreen: React.FC = () => {
 	return (
 		<View>
 			<Text>Welcome to the App!</Text>
-			<RadioGroup
-				options={[
-					{ label: "Family and relationships", value: "family" },
-					{ label: "Career and professional growth", value: "career" },
-					{ label: "Personal development and learning", value: "personal" },
-					{ label: "Health and well-being", value: "health" },
-					{ label: "Financial stability", value: "financial" },
-				]}
-			/>
+			<RadioGroup options={mockData} />
 		</View>
 	);
 };
